refactor(inference): split OpenRouter runtime adapter
