+++ src/lib/inference/openrouter-runtime-adapter-entry.ts
@@ -1,10 +1,7 @@
-#!/usr/bin/env node
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-const {
-  startOpenRouterRuntimeAdapterFromEnv,
-} = require("../dist/lib/inference/openrouter-runtime-adapter");
+import { startOpenRouterRuntimeAdapterFromEnv } from "./openrouter-runtime-adapter-server";
 
 try {
   startOpenRouterRuntimeAdapterFromEnv();
